Rename primitives.tsx to components.tsx and document naming convention
Rename product-card/primitives.tsx to components.tsx. Update AGENTS.md
file organization guidance: use components.tsx for directive-free
sub-components, client-components.tsx for client, server-components.tsx
for server.
